create another main_internal
